[CORE]: Refactored Api handlers into their own independent files
