+++ Cellua/Api/Common/RandomApi.cs
@@ -1,0 +1,16 @@
+﻿using Cellua.Random;
+
+namespace Cellua.Api.Common
+{
+    public class RandomApi
+    {
+        public RandomBool NewRandomBool()
+        {
+            return new RandomBool();
+        }
+        public RandomBool NewRandomBool(int seed)
+        {
+            return new RandomBool(seed);
+        }
+    }
+}
