+++ Cellua/Api/Common/SystemApi.cs
@@ -1,0 +1,12 @@
+﻿using System.Threading;
+
+namespace Cellua.Api.Common
+{
+    public class SystemApi
+    {
+        public static void Sleep(int ms)
+        {
+            Thread.Sleep(ms);
+        }
+    }
+}
